# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.CartAddItemRequest:
    properties:
      product_id:
        type: integer
      quantity:
        minimum: 1
        type: integer
    required:
    - product_id
    - quantity
    type: object
  models.CartUpdateItemRequest:
    properties:
      quantity:
        minimum: 1
        type: integer
    required:
    - quantity
    type: object
...
  models.SwaggerCart:
    description: Cart model for Swagger documentation
    properties:
      created_at:
        example: "2023-01-01T00:00:00Z"
        type: string
      id:
        example: 1
        type: integer
      items:
        items:
          $ref: '#/definitions/models.SwaggerCartItem'
        type: array
      session_id:
        example: session123
        type: string
      status:
        example: active
        type: string
      updated_at:
        example: "2023-01-01T00:00:00Z"
        type: string
      user_id:
        example: 1
        type: integer
    type: object
  models.SwaggerCartItem:
    description: Cart item model for Swagger documentation
    properties:
      cart_id:
        example: 1
        type: integer
      created_at:
        example: "2023-01-01T00:00:00Z"
        type: string
      id:
        example: 1
        type: integer
      product_id:
        example: 1
        type: integer
      quantity:
        example: 2
        type: integer
      updated_at:
        example: "2023-01-01T00:00:00Z"
        type: string
    type: object
...
  response.Response:
    description: Standard API response structure
    properties:
      code:
        example: 200
        type: integer
      data: {}
      error: {}
      message:
        example: Operation successful
        type: string
      status:
        description: '"success" | "error"'
        example: success
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: A TechStore e-commerce API built with Go
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: TechStore API
  version: "1.0"
paths:
...
  /cart:
    delete:
      consumes:
      - application/json
      description: Remove all items from the cart
      produces:
      - application/json
      responses:
        "200":
          description: Cart cleared
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Cart not found
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Clear cart
      tags:
      - cart
    get:
      consumes:
      - application/json
      description: Retrieve the current user's cart
      produces:
      - application/json
      responses:
        "200":
          description: Cart retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.SwaggerCart'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Get cart
      tags:
      - cart
  /cart/items:
    post:
      consumes:
      - application/json
      description: Add a product to the current user's cart
      parameters:
      - description: Cart item data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.CartAddItemRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Item added to cart
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.SwaggerCartItem'
              type: object
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Add item to cart
      tags:
      - cart
  /cart/items/{itemId}:
    delete:
      consumes:
      - application/json
      description: Remove an item from the cart
      parameters:
      - description: Cart Item ID
        in: path
        name: itemId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Item removed from cart
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Cart item not found
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Remove item from cart
      tags:
      - cart
    put:
      consumes:
      - application/json
      description: Update the quantity of an item in the cart
      parameters:
      - description: Cart Item ID
        in: path
        name: itemId
        required: true
        type: string
      - description: Cart item update data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.CartUpdateItemRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Cart item updated
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.SwaggerCartItem'
              type: object
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Cart item not found
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Update cart item
      tags:
      - cart
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"